refactor: address code review feedbacks
