+++ src/sap_cloud_sdk/core/protocol/http/__init__.py
@@ -1,0 +1,15 @@
+"""HTTP client protocol package."""
+
+from sap_cloud_sdk.core.protocol.http.models import (
+    AuthProvider,
+    HttpMethod,
+    XsuaaAuthProvider,
+)
+from sap_cloud_sdk.core.protocol.http.client import HttpClient
+
+__all__ = [
+    "AuthProvider",
+    "HttpClient",
+    "HttpMethod",
+    "XsuaaAuthProvider",
+]
