+++ tests/destination/unit/test_fragment_client.py
@@ -5,7 +5,7 @@ from unittest.mock import Mock
 from requests import Response
 
 from sap_cloud_sdk.destination.fragment_client import FragmentClient
-from sap_cloud_sdk.core._http_client import HttpMethod
+from sap_cloud_sdk.core.protocol.http import HttpMethod
 from sap_cloud_sdk.destination._models import AccessStrategy, Fragment, Label, Level, PatchLabels
 from sap_cloud_sdk.destination.exceptions import (
     DestinationOperationError,
